% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidytime.r
\name{tidytime.default}
\alias{tidytime.default}
\title{Default tidytime method}
\usage{
\method{tidytime}{default}(x, ...)
}
\arguments{
\item{x}{time series object to tidy}

\item{...}{extra arguments}
}
\description{
Default tidytime method
}

