% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidytime.r
\name{tidytime.NULL}
\alias{tidytime.NULL}
\title{Tidier for NULL object}
\usage{
\method{tidytime}{NULL}(x, ...)
}
\arguments{
\item{x}{time series object to tidy}

\item{...}{extra arguments}
}
\description{
Tidier for NULL object
}

