% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidytime.r
\docType{package}
\name{tidytime}
\alias{tidytime}
\alias{tidytime-package}
\title{tidytime: tidying functions for time series data}
\usage{
tidytime(x, ...)
}
\arguments{
\item{x}{time series object to tidy}

\item{...}{extra arguments}
}
\value{
\code{tidytime} returns a \code{data_frame} with one row for each
  observation in each series, with the following columns:

  \item{index}{Index (usually date)}
  \item{series}{Name of the series}
  \item{value}{Value of the observation}
}
\description{
tidytime provides functions for tidying time series data.
\code{tidytime} supports major time series formats including:
\itemize{
 \item builtin time.series (ts)
 \item zoo
 \item xts
 }

This function is an S3 generic.
}
\seealso{
\url{https://cran.r-project.org/web/packages/zoo/index.html}

\url{https://cran.r-project.org/web/packages/xts/index.html}
}

