% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidytime.r
\name{tidytime.ts}
\alias{tidytime.ts}
\title{Tidier for time series object}
\usage{
\method{tidytime}{ts}(x, ...)
}
\arguments{
\item{x}{time series object to tidy}

\item{...}{extra arguments}
}
\value{
data_frame
}
\description{
Tidier for time series object
}
\examples{

}

