% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidytime.r
\name{tidytime.xts}
\alias{tidytime.xts}
\title{Tidier for xts object}
\usage{
\method{tidytime}{xts}(x, ...)
}
\arguments{
\item{x}{time series object to tidy}

\item{...}{extra arguments}
}
\value{
data_frame
}
\description{
Tidier for xts object
}
\examples{
if (require(xts)) {
  data(sample_matrix)
  sample.xts <- as.xts(sample_matrix)
  tidy(sample.xts)
}

}

