% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidytime.r
\name{tidytime.zoo}
\alias{tidytime.zoo}
\title{Tidier for zoo object}
\usage{
\method{tidytime}{zoo}(x, ...)
}
\arguments{
\item{x}{time series object to tidy}

\item{...}{extra arguments}
}
\value{
data_frame
}
\description{
Tidier for zoo object
}
\examples{
x.Date <- as.Date("2003-02-01") + c(1, 3, 7, 9, 14) - 1
x <- zoo(rnorm(5), x.Date)
tidytime(x)

}

